pyproject.toml
[project]
name = "secwatcher"
version = "0.1.0"
description = "Autonomous GitHub repository security scanner: secrets, Actions, branch protection, dep vuln. Slack-delivered."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Oleg Strutsovski", email = "cyrillical@gmail.com" }]
keywords = ["security", "github", "secret-scanning", "appsec", "gitleaks", "trufflehog"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "PyGithub>=2.3.0",
    "PyJWT[crypto]>=2.8.0",
    "httpx>=0.27.0",
    "slack-sdk>=3.27.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
secwatcher = "secwatcher.cli:app"
reposentry = "secwatcher.cli:app"

[project.urls]
Homepage = "https://github.com/cyrillical00/secwatcher"
Issues = "https://github.com/cyrillical00/secwatcher/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["secwatcher"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S", "PL", "RUF"]
ignore = [
    "S603",  # subprocess: trufflehog/gitleaks invocation is intentional
    "S607",  # partial executable path: binaries resolved via PATH
    "PLR0913",  # too many arguments: dataclass-style constructors
    "PLR2004",  # HTTP status codes and similar literals are clearer inline
    "UP042",  # str-Enum subclassing is more compatible than StrEnum
    "PLC0415",  # late imports needed in orchestrator factory to break cycles
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "PLR2004"]
"secwatcher/cli.py" = ["B008"]  # typer.Option(...) is the framework pattern

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
